Paso 1 de ruta mapa 3

+++ game1/src/clasesConSprite/Pompa.hpp
@@ -68,7 +68,7 @@ public:
 	static const int EXPLOTA = 4;
 	static const int NUM_FOTOGRAMAS = 3;
 
-	static uint32_t ID_MAPA;	//Debe actualizarse cada vez que se cambie el mapa
+	static uint8_t ID_MAPA;	//Debe actualizarse cada vez que se cambie el mapa
 	static uint32_t ID_GLOBAL;
 	int ID = -1;
 	static void init();
